@@ -1,50 +1,94 @@
-import React, { Component } from "react";
-import { Switch, Route, Link } from "react-router-dom";
-import "bootstrap/dist/css/bootstrap.min.css";
-import "./App.css";
-
-import AddTutorial from "./components/add-tutorial.component";
-import TutorialsList from "./components/tutorials-list.component";
-import AddCustomer from "./components/add-customer.component";
-
-class App extends Component {
-  render() {
-    return (
-      <div>
-        <nav className="navbar navbar-expand navbar-dark bg-dark">
-          <a href="/tutorials" className="navbar-brand">
-            bezKoder
-          </a>
-          <div className="navbar-nav mr-auto">
-            <li className="nav-item">
-              <Link to={"/tutorials"} className="nav-link">
-                Tutorials
-              </Link>
-            </li>
-            <li className="nav-item">
-              <Link to={"/add"} className="nav-link">
-                Add
-              </Link>
-            </li>
-            <li className="nav-item">
-              <Link to={"/register"} className="nav-link">
-                Register
-              </Link>
-            </li>
-          </div>
-        </nav>
-
-        <div className="container mt-3">
-          <h2>React Firebase Database CRUD</h2>
-          <Switch>
-            <Route exact path={["/", "/tutorials"]} component={TutorialsList} />
-            <Route exact path="/add" component={AddTutorial} />
-            <Route exact path="/register" component={AddCustomer} />
-          </Switch>
-        </div>
-      </div>
-    );
+import { useState, useEffect } from 'react';
+import './App.css';
+import Form from './components/common/form';
+import Home from './components/home.component';
+import {
+  Routes,
+  Route,
+  useNavigate
+} from "react-router-dom";
+import { app } from './firebase-db-config';
+import { getAuth, signInWithEmailAndPassword, createUserWithEmailAndPassword } from 'firebase/auth';
+import { ToastContainer, toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+function App() {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  let navigate = useNavigate();
+  const handleAction = (id) => {
+    const authentication = getAuth();
+    if (id === 1) {
+      signInWithEmailAndPassword(authentication, email, password)
+          .then((response) => {
+            navigate('/home')
+            sessionStorage.setItem('Auth Token', response._tokenResponse.refreshToken)
+          })
+          .catch((error) => {
+            console.log(error.code)
+            if (error.code === 'auth/wrong-password') {
+              toast.error('Please check the Password');
+            }
+            if (error.code === 'auth/user-not-found') {
+              toast.error('Please check the Email');
+            }
+          })
+    }
+    if (id === 2) {
+      createUserWithEmailAndPassword(authentication, email, password)
+          .then((response) => {
+            navigate('/home')
+            sessionStorage.setItem('Auth Token', response._tokenResponse.refreshToken)
+          })
+          .catch((error) => {
+            if (error.code === 'auth/email-already-in-use') {
+              toast.error('Email Already in Use');
+            }
+          })
+    }
   }
+
+  useEffect(() => {
+    let authToken = sessionStorage.getItem('Auth Token')
+
+    if (authToken) {
+      navigate('/home')
+    }
+  }, [])
+  return (
+      <div className="App">
+        <>
+          <ToastContainer />
+          <Routes>
+            <Route
+                path='/login'
+                element={
+                  <Form
+                      title="Login"
+                      setEmail={setEmail}
+                      setPassword={setPassword}
+                      handleAction={() => handleAction(1)}
+                  />}
+            />
+            <Route
+                path='/register'
+                element={
+                  <Form
+                      title="Register"
+                      setEmail={setEmail}
+                      setPassword={setPassword}
+                      handleAction={() => handleAction(2)}
+                  />}
+            />
+
+            <Route
+                path='/home'
+                element={
+                  <Home />}
+            />
+          </Routes>
+        </>
+      </div>
+  );
 }
 
 export default App;
